#213 Refactor: divided into 2 types of errors: errors DOM and errors API
