+++ typescript/big-practice/src/scripts/controllers/noteController.js
@@ -1,4 +1,4 @@
-import { renderPopupError } from '../utils/handleError';
+import { renderPopupError } from '../utils/errorsDOM';
 import LoadingPage from '../utils/loadingPage';
 
 /**
